Rename createConfigProvider to register
